extract method for searching for main class
